Google login

+++ app/Repositories/Users.php
@@ -101,8 +101,10 @@ class Users extends Repository
             $instance->offsetUnset('password');
         }
 
-        $instance->roles()->detach();
-        $instance->assignRole($attributes['role']);
+        if (isset($attributes['role'])){
+            $instance->roles()->detach();
+            $instance->assignRole($attributes['role']);
+        }
 
         if (is_null($instance->getRememberToken())) {
             $instance->setRememberToken(Str::random(60));
@@ -133,5 +135,17 @@ class Users extends Repository
         }
 
         return $this->create($attributes);
+    }
+
+    /**
+     * Finds a user by its email attribute.
+     *
+     * @param string $email
+     * @param array $options
+     * @return User|null
+     */
+    public function findByEmail(string $email, array $options = []): ?User
+    {
+        return $this->findBy(['email' => $email], $options);
     }
 }
